* speex data is now encoded (but still not sent)  * fixed a compiler warning in config debug level parsing and push to talk
git-svn-id: http://svn.mangler.org/mangler/trunk@148 f654edcb-a5fb-4f28-a0e7-68c327795dae

+++ src/manglersettings.cpp
@@ -185,7 +185,7 @@ void ManglerSettings::initSettings(void) {/*{{{*/
     checkbutton->set_active(config.lv3_debuglevel & V3_DEBUG_SOCKET ? 1 : 0);
 
     builder->get_widget("debugNotice", checkbutton);
-    checkbutton->set_active(config.lv3_debuglevel & V3_DEBUG_NOTICE ? 1 : 0);
+    checkbutton->set_active(config.lv3_debuglevel & (uint32_t)V3_DEBUG_NOTICE ? 1 : 0);
 
     builder->get_widget("debugInfo", checkbutton);
     checkbutton->set_active(config.lv3_debuglevel & V3_DEBUG_INFO ? 1 : 0);
